modified Drinking Song function
I thought about making your Drinking Song feature a little smoother.
First only one groupaction is needed. The function itself should check if there is a hobbit or not. So i made a def "drinkingSong" which will be triggered by Handoption.
This def checks the cards on the table and hopefully found an unique Hobbit controlled by me, then additionalCards will be 1 otherwise 0.
After that the more allround def "exchangeCardsInHand" is called with the value of additionalCards.
Secondly i changed the notify in "exchangeCardsInHand" so it fits to the Handsize after drawing (shared.HandSize is always 6)